Models

+++ app/controllers/User.java
@@ -2,7 +2,7 @@ package controllers;
 
 import play.data.*;
 import play.mvc.*;
-import views.html.users.*;
+import views.html.user.*;
 import java.util.*;
 import models.*;
 import play.db.ebean.*;
@@ -10,15 +10,14 @@ import javax.persistence.*;
 import org.mindrot.jbcrypt.BCrypt;
 
 @Security.Authenticated(Secured.class)
-public class Users extends Controller {
+public class User extends Controller {
 
-	private static final Form<User> userForm = Form.form(User.class);
-	private static final User current_user = User.findByEmail(request().username());
+	private static final Form<Users> userForm = Form.form(Users.class);
 
 	public static Result index() {
 		if (Secured.isAdmin()) {
-			List<User> users = User.findAll();
-			return ok(index.render(users, current_user));
+			List<Users> users = Users.findAll();
+			return ok(index.render(users, Users.findByEmail(request().username())));
 		} else {
 			return forbidden("You don't have permission to access on this server");
 		}
@@ -26,7 +25,7 @@ public class Users extends Controller {
 
 	public static Result newRecord() {
 		if (Secured.isAdmin()) {
-			return ok(news.render(userForm, current_user));
+			return ok(news.render(userForm, Users.findByEmail(request().username())));
 		} else {
 			return forbidden("You don't have permission to access on this server");
 		}
@@ -34,7 +33,7 @@ public class Users extends Controller {
 
 	public static Result create() {
 		if (Secured.isAdmin()) {
-			Form<User> boundForm = userForm.bindFromRequest();
+			Form<Users> boundForm = userForm.bindFromRequest();
 
 			String password_confirmation = boundForm.field("password_confirmation").valueOr("");
 			String password = boundForm.field("password").valueOr("");
@@ -45,24 +44,24 @@ public class Users extends Controller {
 
 			if (boundForm.hasErrors()) {
 				flash("error", "Please correct the form below.");
-				return badRequest(news.render(boundForm, current_user));
+				return badRequest(news.render(boundForm, Users.findByEmail(request().username())));
 			}
 
-			User user = boundForm.get();
+			Users user = boundForm.get();
 			user.password = BCrypt.hashpw(user.password, BCrypt.gensalt());
 			user.save();
 			flash("success", String.format("Successfully added user."));
-			return redirect(routes.Users.index());
+			return redirect(routes.User.index());
 		} else {
 			return forbidden("You don't have permission to access on this server");
 		}
 	}
 
 	public static Result update(Long id) {
-		Form<User> boundForm = userForm.bindFromRequest();
+		Form<Users> boundForm = userForm.bindFromRequest();
 
-		User user = User.findById(id);
-		if (user != current_user) {
+		Users user = Users.findById(id);
+		if (!user.email.equals(request().username()) && !Secured.isAdmin()) {
 			return forbidden("You don't have permission to access on this server");
 		}
 		String password = boundForm.field("password").valueOr("");
@@ -80,7 +79,7 @@ public class Users extends Controller {
 
 		if (boundForm.hasErrors()) {
 			flash("error", "Please correct the form below.");
-			return badRequest(edit.render(boundForm, user, current_user));
+			return badRequest(edit.render(boundForm, user, Users.findByEmail(request().username())));
 		}
 		user = boundForm.get();
 		if (!new_password.isEmpty()) {
@@ -91,32 +90,38 @@ public class Users extends Controller {
 		user.update();
 		flash("success", String.format("Successfully updated user."));
 		if (Secured.isAdmin()) {
-			return redirect(routes.Users.index());
+			return redirect(routes.User.index());
 		} else {
 			return redirect(routes.Employs.index());
 		}
 	}
 
 	public static Result edit(Long id) {
-		final User user = User.findById(id);
+		final Users user = Users.findById(id);
+		Users current_user = Users.findByEmail(request().username());
 		if (user == null) {
 			return notFound(String.format("User does not exist."));
 		}
-		if (user.email != request().username() && Secured.isAdmin() == false) {
+		if (!user.equals(current_user)) {
 			return forbidden("You don't have permission to access on this server");
 		}
-		Form<User> filledForm = userForm.fill(user);
+		Form<Users> filledForm = userForm.fill(user);
 		return ok(edit.render(filledForm, user, current_user));
 	}
 
 	public static Result destroy(Long id) {
 		if (Secured.isAdmin()) {
-			final User user = User.findById(id);
+			final Users user = Users.findById(id);
+			Users current_user = Users.findByEmail(request().username());
 			if (user == null) {
 				return notFound(String.format("User does not exists."));
 			}
-			user.delete();
-			return redirect(routes.Users.index());
+			if (user.equals(current_user)) {
+				flash("error", String.format("You aren't delete yoursefl."));
+			} else {
+				user.delete();
+			}
+			return redirect(routes.User.index());
 		} else {
 			return forbidden("You don't have permission to access on this server");
 		}
